requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
albumentations==1.3.0
pyyaml==6.0
tqdm==4.66.1
pytest==7.4.0
scikit-image==0.21.0

# MLOps (opcional)
mlflow==2.7.0
wandb==0.15.0
dvc==3.0.0

# Desarrollo
black==23.9.1
flake8==6.1.0
isort==5.12.0
